# Polygon `move()/move_ip()` optimization